+++ src/main/java/algorithm/GRAPH/DFS_BFS/BOJ14500_테트로미노.java
@@ -1,4 +1,4 @@
-package algorithm.DFS_BFS;
+package algorithm.GRAPH.DFS_BFS;
 
 import java.io.BufferedReader;
 import java.io.IOException;
