+++ src/main/java/algorithm/GRAPH/DFS_BFS/BOJ15652_N과M4.java
@@ -1,4 +1,4 @@
-package algorithm.DFS_BFS;
+package algorithm.GRAPH.DFS_BFS;
 
 import java.util.ArrayList;
 import java.util.Scanner;
