+++ src/main/java/algorithm/GRAPH/DFS_BFS/BOJ21736_헌내기는친구가필요해.java
@@ -1,4 +1,4 @@
-package algorithm.DFS_BFS;
+package algorithm.GRAPH.DFS_BFS;
 
 import java.io.BufferedReader;
 import java.io.IOException;
