+++ src/main/java/algorithm/GRAPH/DFS_BFS/Programmers_퍼즐조각채우기.java
@@ -1,4 +1,4 @@
-package algorithm.DFS_BFS;
+package algorithm.GRAPH.DFS_BFS;
 
 import java.util.*;
