+++ src/test/java/algorithm/DFS_BFS/Programmers_퍼즐조각채우기Test.java
@@ -1,5 +1,6 @@
 package algorithm.DFS_BFS;
 
+import algorithm.GRAPH.DFS_BFS.Programmers_퍼즐조각채우기;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
