+++ src/test/java/algorithm/GRAPH/Programmers_순위Test.java
@@ -1,9 +1,9 @@
 package algorithm.GRAPH;
 
+import algorithm.GRAPH.FloydWarshall.Programmers_순위;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 class Programmers_순위Test {
     Programmers_순위 sol = new Programmers_순위();
